requirements/prod.txt
alabaster==0.7.10
asn1crypto==0.22.0
Babel==2.4.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
coverage==4.4.1
coveralls==1.1
cryptography==2.0
Django==1.11.3
django-cors-headers==2.1.0
django-robots==3.0
docopt==0.6.2
docutils==0.13.1
idna==2.5
imagesize==0.7.1
Jinja2==2.9.6
MarkupSafe==1.0
profilehooks==1.9.0
pycparser==2.18
pygit2==0.26.0
Pygments==2.2.0
PyJWT==1.5.2
pytz==2017.2
requests==2.18.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
tokenlib==0.3.1
urllib3==1.21.1
